do refactoring

+++ modules/stream/src/test/java/com/nautilus_technologies/tsubakuro/channel/stream/sql/SessionWireTest.java
@@ -21,7 +21,7 @@ import com.nautilus_technologies.tsubakuro.util.ByteBufferInputStream;
 import com.tsurugidb.jogasaki.proto.SqlResponse;
 
 class SessionWireTest {
-    static final long SERVICE_ID_SQL = 3;
+    static final int SERVICE_ID_SQL = 3;
     private static final String HOST = "localhost";
     private static final int PORT = 12344;
